+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -74,7 +74,10 @@ from ska_mid_dish_manager.models.dish_enums import (
 from ska_mid_dish_manager.models.dish_mode_model import DishModeModel
 from ska_mid_dish_manager.models.dish_state_transition import StateTransition
 from ska_mid_dish_manager.utils.action_helpers import report_task_progress, update_task_status
-from ska_mid_dish_manager.utils.decorators import check_communicating
+from ska_mid_dish_manager.utils.decorators import (
+    check_communicating,
+    last_command_failure_decorator,
+)
 from ska_mid_dish_manager.utils.input_validation import (
     ConfigureBandValidationError,
     validate_configure_band_input,
@@ -197,6 +200,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             windgust=-1,
             lastcommandedmode=("0.0", ""),
             lastcommandinvoked=("0.0", ""),
+            lastcommandfailure=("0.0", "", ""),
             dscctrlstate=DscCtrlState.NO_AUTHORITY,
             rfcmplllock=B5dcPllState.NOT_LOCKED,
             rfcmhattenuation=0.0,
@@ -493,6 +497,25 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
     # --------------
     # Helper methods
     # --------------
+    def publish_last_command_failure(self, command: str, reason: str | None = None) -> None:
+        """Update the last command failure attribute.
+
+        Constructs a `lastcommandfailure` record for a failed command and
+        publishes it via the component state callback.
+
+        :param command: The name of the command that failed.
+        :param status: The status returned by the command.
+        :param reason: The reason or failure message returned by the command.
+        """
+        reason = reason or ""
+        failure = (
+            str(time.time()),
+            command,
+            reason,
+        )
+
+        self._update_component_state(lastcommandfailure=failure)
+
     def get_current_tai_offset_from_dsc_with_manual_fallback(self) -> float:
         """Try and get the TAI offset from the DSManager device.
         Or calulate it manually if that fails.
@@ -1263,6 +1286,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             for component_manager in self.sub_component_managers.values():
                 component_manager.stop_communicating()
 
+    @last_command_failure_decorator
     def reset_subservient_dev_connections(self, device_name: str) -> Tuple[ResultCode, str]:
         """Reset connections between dish manager and subservient devices."""
         if not device_name:
@@ -1385,6 +1409,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         except tango.DevFailed:
             raise
 
+    @last_command_failure_decorator
     def track_load_table(
         self, sequence_length: int, table: list[float], load_mode: TrackTableLoadMode
     ) -> Tuple[TaskStatus, str]:
@@ -1396,6 +1421,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return task_status, msg
 
     @check_communicating
+    @last_command_failure_decorator
     def set_standby_lp_mode(
         self, task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1414,6 +1440,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def set_standby_fp_mode(
         self, task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1432,6 +1459,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def set_maintenance_mode(
         self, task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1451,6 +1479,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def track_cmd(self, task_callback: Optional[Callable] = None) -> Tuple[TaskStatus, str]:
         """Track the commanded pointing position."""
 
@@ -1461,6 +1490,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     "Track command is allowed for dishMode OPERATE"
                 )
                 report_task_progress(msg, self._command_progress_callback)
+                self.publish_last_command_failure("track", msg)
                 return False
             if self.component_state["pointingstate"] != PointingState.READY:
                 msg = (
@@ -1468,6 +1498,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     "Track command is allowed for pointingState READY"
                 )
                 report_task_progress(msg, self._command_progress_callback)
+                self.publish_last_command_failure("track", msg)
                 return False
             return True
 
@@ -1479,6 +1510,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def track_stop_cmd(self, task_callback: Optional[Callable] = None) -> Tuple[TaskStatus, str]:
         """Stop tracking."""
 
@@ -1489,6 +1521,11 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 PointingState.TRACK,
                 PointingState.SLEW,
             ]:
+                self.publish_last_command_failure(
+                    "track_stop",
+                    "Command not allowed. DishMode is not OPERATE and"
+                    " PointingState is not SLEW/TRACK.",
+                )
                 return False
             return True
 
@@ -1500,6 +1537,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def set_operate_mode(self, task_callback: Optional[Callable] = None) -> Tuple[TaskStatus, str]:
         """Fanout commands to subdevices to set operate mode.
 
@@ -1518,6 +1556,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     "Set operate mode command is allowed for dishMode STANDBY_FP"
                 )
                 report_task_progress(msg, self._command_progress_callback)
+                self.publish_last_command_failure("set_operate", msg)
                 return False
             return True
 
@@ -1532,6 +1571,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def configure_band_cmd(
         self, band: Band, synchronise: bool, task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1574,6 +1614,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def configure_band_six_cmd(
         self,
         task_callback: Optional[Callable] = None,
@@ -1596,6 +1637,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return status, response
 
     @check_communicating
+    @last_command_failure_decorator
     def configure_band_with_json(
         self,
         data: str,
@@ -1666,6 +1708,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         )
         self._update_component_state(dishmode=new_dish_mode)
 
+    @last_command_failure_decorator
     def set_stow_mode(self, task_callback: Optional[Callable] = None) -> Tuple[TaskStatus, str]:
         """Transition the dish to STOW mode."""
         ds_cm = self.sub_component_managers["DS"]
@@ -1687,6 +1730,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return TaskStatus.COMPLETED, "Stow called, monitor dishmode for LRC completed"
 
     @check_communicating
+    @last_command_failure_decorator
     def slew(
         self, values: list[float], task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1705,6 +1749,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     "Slew command is allowed for dishMode OPERATE"
                 )
                 report_task_progress(msg, self._command_progress_callback)
+                self.publish_last_command_failure("slew", msg)
                 return False
             if self.component_state["pointingstate"] != PointingState.READY:
                 msg = (
@@ -1712,6 +1757,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     "Slew command is allowed for pointingState READY"
                 )
                 report_task_progress(msg, self._command_progress_callback)
+                self.publish_last_command_failure("slew", msg)
                 return False
             return True
 
@@ -1726,6 +1772,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         )
         return status, response
 
+    @last_command_failure_decorator
     def scan(
         self, scanid: str, task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1745,6 +1792,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             task_callback, status=TaskStatus.COMPLETED, result=(ResultCode.OK, "Scan completed")
         )
 
+    @last_command_failure_decorator
     def end_scan(self, task_callback: Optional[Callable] = None) -> Tuple[TaskStatus, str]:
         """Clear the scanid."""
         status, response = self.submit_task(self._end_scan, args=[], task_callback=task_callback)
@@ -1761,6 +1809,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         )
 
     @check_communicating
+    @last_command_failure_decorator
     def track_load_static_off(
         self, values: list[float], task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]:
@@ -1784,6 +1833,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         )
         return status, response
 
+    @last_command_failure_decorator
     def set_kvalue(self, k_value: float) -> Tuple[ResultCode, str]:
         """Set the k-value on the SPFRx.
         Note that it will only take effect after
@@ -1795,6 +1845,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             return (ResultCode.FAILED, msg)
         return (ResultCode.OK, msg)
 
+    @last_command_failure_decorator
     def apply_pointing_model(self, json_object) -> Tuple[ResultCode, str]:
         # pylint: disable=R0911
         """Updates a band's coefficient parameters with a given JSON input.
@@ -2079,6 +2130,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return (ResultCode.OK, "Successfully updated dscPowerLimitKw on DS")
 
     @check_communicating
+    @last_command_failure_decorator
     def reset_track_table(self) -> Tuple[ResultCode, List[float] | str]:
         """Reset the track table."""
         self.logger.debug("Resetting the programTrackTable")
@@ -2099,6 +2151,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             return ResultCode.FAILED, msg
         return ResultCode.OK, reset_point
 
+    @last_command_failure_decorator
     def set_h_pol_attenuation(self, value: int) -> Tuple[ResultCode, str]:
         """Update HPolAttenuation on B5DC."""
         if not self._check_b5dc_active():
@@ -2113,6 +2166,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             return (ResultCode.FAILED, msg)
         return (ResultCode.OK, "Successfully updated HPolAttenuation on the B5DC proxy.")
 
+    @last_command_failure_decorator
     def set_v_pol_attenuation(self, value: int) -> Tuple[ResultCode, str]:
         """Update VPolAttenuation on B5DC."""
         if not self._check_b5dc_active():
@@ -2127,6 +2181,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             return (ResultCode.FAILED, msg)
         return (ResultCode.OK, "Successfully updated VPolAttenuation on the B5DC proxy.")
 
+    @last_command_failure_decorator
     def set_frequency(self, value: int) -> Tuple[ResultCode, str]:
         """Update Frequency on B5DC."""
         if not self._check_b5dc_active():
@@ -2142,6 +2197,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         return (ResultCode.OK, "Successfully updated Frequency on the B5DC proxy.")
 
     @check_communicating
+    @last_command_failure_decorator
     def interlock_acknowledge(
         self, task_callback: Optional[Callable] = None
     ) -> Tuple[TaskStatus, str]: